Fix(Platform) - Good enough + Fix(Event) - axis_y

+++ include/game/systems/world/camera_follow.h
@@ -38,7 +38,7 @@ namespace Game::World
         }
 
         // Cam follow
-        camera_pos.x += (camera_target.x - camera_pos.x) * CAMERA_SPEED * get_delta_time();
-        camera_pos.y += (camera_target.y - camera_pos.y) * CAMERA_SPEED * get_delta_time();
+        camera_pos.x += (camera_target.x - camera_pos.x) * CAMERA_SPEED * (float)get_delta_time();
+        camera_pos.y += (camera_target.y - camera_pos.y) * CAMERA_SPEED * (float)get_delta_time();
     }
 } // namespace Game::Overview
